Fix some rupees being counted twice
The fact that some checks are in the logic more than once was news to me.
Still not sure whether the randomizer actually properly handles this
instead of silently dropping some items on a check twice, but we now
handle the base case in the map tracker in order to show the right checks
being available.